+++ moonlight-clock@torchipeppo/files/moonlight-clock@torchipeppo/5.4/color_scheme.js
@@ -1,0 +1,1 @@
+../3.4/color_scheme.js