+++ yarr@jtoberling/CHANGELOG.md
@@ -5,9 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+<<<<<<< HEAD
 ## [1.4.2] - 2025-03-30
 - Performance optimization
 
+=======
+>>>>>>> refs/remotes/origin/master
 ## [1.4.1] - 2025-02-09
 
 - Add Density feature, fixin password dialog width